task 0 attemp

+++ 0x00-python_variable_annotations/0-add.py
@@ -2,6 +2,6 @@
 """ adds two integers and return a float"""
 
 
-def add(a: int, b: int)-> float:
+def add(a: float, b: float)-> float:
     """ all integers are annotated, and the return is float"""
     return a + b
